fix: 위치 필드에서 scale 속성에 의한 오류 수정
- Double -> BigDecimal 타입으로 변경

+++ src/main/java/com/uhdyl/backend/user/domain/User.java
@@ -4,6 +4,7 @@ import com.uhdyl.backend.global.base.BaseEntity;
 import com.uhdyl.backend.global.oauth.user.OAuth2Provider;
 import com.uhdyl.backend.review.domain.Review;
 import jakarta.persistence.*;
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -40,10 +41,10 @@ public class User extends BaseEntity {
     private String publicId;
 
     @Column(precision = 9, scale = 6)
-    private Double location_x;
+    private BigDecimal location_x;
 
     @Column(precision = 9, scale = 6)
-    private Double location_y;
+    private BigDecimal location_y;
   
     @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<Review> reviews = new ArrayList<>();
@@ -63,7 +64,7 @@ public class User extends BaseEntity {
         this.publicId = null;
     }
 
-    public void updateLocation(Double location_x, Double location_y){
+    public void updateLocation(BigDecimal location_x, BigDecimal location_y){
         this.location_x = location_x;
         this.location_y = location_y;
     }
